Move package to the right location
