+++ cmd/worker/internal/gitserver/background.go
@@ -1,5 +1,5 @@
-// Package background contains all the background jobs related to gitserver
+// Package gitserver contains all the background jobs related to gitserver
 // that are running in the worker service.
 // Components of this package implement the worker service background job interface
 // which is documented here https://docs.sourcegraph.com/dev/background-information/workers.
-package background
+package gitserver
