+++ cmd/worker/internal/gitserver/local_clone_test.go
@@ -1,4 +1,4 @@
-package background
+package gitserver
 
 import (
 	"context"
